### 2024-10-22 #7397 @kthyng
👋 @danStich - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it.
