+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-10-22T15:37Z
+last_built: 2024-10-22T21:12Z
 urls:
   reference: https://danstich.github.io/dia/reference
   article: https://danstich.github.io/dia/articles
