+++ paper.md
@@ -14,18 +14,18 @@ authors:
   - name: Daniel S. Stich
     orcid: "0000-0002-8946-1115"
     corresponding: true
-    affiliation: 1
+    affiliation: "1"
   - name: Julie L. Nieland
-    affiliation: 2
+    affiliation: "2"
   - name: Timothy F. Sheehan
-    affiliation: 2
+    affiliation: "2"
 affiliations:
 - name: Biology Department and Biological Field Station, State University of New York at Oneonta, NY 13280 USA
   index: 1
 - name: National Oceanic and Atmospheric Administration, National Marine Fisheries Service, Northeast Fisheries Science Center, Woods Hole, MA 02543 USA
   index: 2
+date: 22 October 2024
 bibliography: paper.bib
-date: 10 September 2024
 ---
 
 # Summary
