+++ paper.bib
@@ -48,7 +48,7 @@
 	VOLUME = {8},
 	YEAR = {2021},
 	URL = {https://www.frontiersin.org/journals/marine-science/articles/10.3389/fmars.2021.734213},
-	DOI = {10.3389/fmars.2021.734213},
+	doi = {10.3389/fmars.2021.734213},
 	ISSN = {2296-7745}
 }
 
@@ -72,7 +72,7 @@
     month = {05},
     issn = {1054-3139},
     doi = {10.1093/icesjms/fsv083},
-    url = {https://doi.org/10.1093/icesjms/fsv083},
+    url = {10.1093/icesjms/fsv083},
     eprint = {https://academic.oup.com/icesjms/article-pdf/72/8/2423/31226690/fsv083.pdf}
 }
 
