### 2024-10-23 #7397 @danStich
Sounds great, thank you. I can find the 'skipped' DOIs from the reference check in the meantime and start putting together a list of potential reviewers.
