### 2024-10-24 #7397 @kthyng
@danStich This looks ready to go though an example plot in your docs would be great — something to look at, a little more of a tutorial sort of thing. Just a suggestion.
