### 2024-10-24 #7397 @danStich
@kthyng Thank you, I can add an example plot. I think we may have had one in an earlier internal draft as well. I have another MS in revision that has a short turnaround time so it may be Monday before I address the items I added for myself to the list below.

- [ ] Skip DOIs
- [ ] 5 Reviewers from database listed above or extended, non-conflicted network
- [ ] example plot
