### 2024-10-28 082e235
update bibtex file with available missing DOIs
