+++ paper.bib
@@ -58,7 +58,8 @@
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/danStich/dia}
+  url = {https://github.com/danStich/dia},
+  doi = {10.5281/zenodo.13376045}
 }
 
 @article{nieland:2015,
@@ -96,7 +97,7 @@
   volume = {10},
   number = {2},
   pages = {236-254},
-  doi = {https://doi.org/10.1002/mcf2.10021},
+  doi = {10.1002/mcf2.10021},
   url = {https://afspubs.onlinelibrary.wiley.com/doi/abs/10.1002/mcf2.10021},
   eprint = {https://afspubs.onlinelibrary.wiley.com/doi/pdf/10.1002/mcf2.10021},
   year = {2018}
@@ -108,6 +109,7 @@
   year = {2022},
   note = {R package version 1.7.3},
   url = {https://CRAN.R-project.org/package=shiny},
+  doi = {10.32614/CRAN.package.shiny}
 }
  
 @Manual{R,
@@ -116,7 +118,7 @@
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2022},
-  url = {https://www.R-project.org/},
+  url = {https://www.R-project.org/}
 }  
   
 @Book{ggplot2,
@@ -126,6 +128,7 @@
   year = {2016},
   isbn = {978-3-319-24277-4},
   url = {https://ggplot2.tidyverse.org},
+  doi = {10.32614/CRAN.package.ggplot2}
 }
 
 @Article{tidyverse,
