### 2024-10-28 4973e7b
add example figure
