+++ paper.md
@@ -38,7 +38,9 @@ Anadromous (sea-run) fishes such as Atlantic salmon *Salmo salar* have been grea
 
 We created `dia` for use by fisheries researchers, managers, and practitioners interested in understanding population dynamics of intensively managed endangered Atlantic salmon in the USA. The R package maintains the core routines from the original closed-source version of the model by replicating spreadsheet-based calculations, and incorporates original data and parameter sets as built-in objects that serve as default values for arguments of the primary user-facing functions. 
 
-The two primary user-facing functions within the `dia` package are the `run_dia()` and `run_dia_shiny()`, which provide redundant interfaces for using Dam Impact Analysis models in different ways. The `run_dia()` function provides an extensible interface to DIA that can be used for long-run simulation or decision-optimization studies, and allows incorporation of user-specified data sets such as flow-correlated survival probabilities at dams and in free-flowing river reaches, marine survival and other life-history inputs, or fish-stocking data. The `run_dia_shiny()` function deploys a graphical user interface using the `shiny` package (@shiny) that is less extensible but more easily used by fishery managers and practitioners who may be less familiar with programming. `run_dia_shiny()` also includes exportable results from simulation models including `.csv` or other flat-file formats and default plots through the `ggplot2` R package [@ggplot2; @tidyverse]. Both can be deployed on networked servers as any other R or `shiny` application to improve accessibility or facilitate use on high performance computers for large simulations. The GitHub repository [@dia] includes additional instructions for installation and a variety of potential uses of `run_dia()` and `run_dia_shiny()` in addition to shorter examples in the package help files. While implementation is currently limited to the Penobscot River as a priority conservation water in the USA, it provides increased flexibility for extending this approach to Atlantic salmon and other sea-run fish globally. 
+The two primary user-facing functions within the `dia` package are the `run_dia()` and `run_dia_shiny()`, which provide redundant interfaces for using Dam Impact Analysis models in different ways. The `run_dia()` function provides an extensible interface to DIA that can be used for long-run simulation or decision-optimization studies, and allows incorporation of user-specified data sets such as flow-correlated survival probabilities at dams and in free-flowing river reaches, marine survival and other life-history inputs, or fish-stocking data. The `run_dia_shiny()` function deploys a graphical user interface using the `shiny` package (@shiny) that is less extensible but more easily used by fishery managers and practitioners who may be less familiar with programming. `run_dia_shiny()` also includes exportable results from simulation models including `.csv` or other flat-file formats and default plots through the `ggplot2` R package [@ggplot2; @tidyverse; Figure 1]. Both can be deployed on networked servers as any other R or `shiny` application to improve accessibility or facilitate use on high performance computers for large simulations. The GitHub repository [@dia] includes additional instructions for installation and a variety of potential uses of `run_dia()` and `run_dia_shiny()` in addition to shorter examples in the package help files. While implementation is currently limited to the Penobscot River as a priority conservation water in the USA, it provides increased flexibility for extending this approach to Atlantic salmon and other sea-run fish globally. 
+
+![Figure1. Example graphical outputs using the default argument values (@nieland:2020) for `run_dia()` to run 10,000 simulations, showing (a) the number of two-sea-winter adult females of hatchery or wild origin returning to the watershed during each generation, and (b) the number of those fish returning to each production unit within the watershed after 15 generations.](Figure1.jpg)
 
 # Acknowledgements
[…]
