### 2024-10-28 #7397 @danStich
Checking that PDF compiles following updated DOIs and addition of example figure (Figure 1).
