### 2024-11-08 #7397 @rmk118
Hi @cheginit! I would be happy to edit this submission. This will be my first JOSS review, so I apologize in advance if I am not the smoothest at navigating the logistics of reviewing on GitHub.
