### 2024-11-09 #7397 @cheginit
@rmk118 Thanks for agreeing to review this submission and welcome to the JOSS community! I will provide detailed instructions for the review, and you can also read JOSS's [guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information.

Once I find another reviewer, I will open a new issue and ping you with instructions to being the review. Thanks!
