### 2024-11-13 #7475 @editorialbot
[REVIEW]: dia: An R package for the National Oceanic and Atmospheric Administration dam impact analysis
**Submitting author:** @danStich (<a href="http://orcid.org/0000-0002-8946-1115">Daniel Stich</a>)
**Repository:** https://github.com/danStich/dia
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.1
