### 2024-11-13 #7475 @rmk118
Here is my initial review of the submission! If needed, I would be happy to open specific issues/PRs on the software repository for each point I mentioned below, but many of them are just minor points and I was not sure separate issues were warranted. I hope this is helpful, and let me know if you have any questions or concerns!

- - -
This submission provides an open-source tool to replace a previous Excel-based implementation of the DIA. As the Excel version has been mentioned in several scientific articles and federal reference documents, it seems likely that at a minimum, this package/paper will be cited by researchers who previously used the Excel model. The major limitation of this package is the extreme specificity to Atlantic salmon in the Penobscot River. While some of the helper functions (like `make_eggs_per_female()`) are relatively generalizable, arguments to the main package functions include parameters for specific dams on the Penobscot, and the package functionality is highly dependent on the built-in Penobscot River datasets. However, the Penobscot is a priority conservation area intensely studied by fisheries scientists, so there are still many researchers for whom this would be of interest. Furthermore, the relatively niche appeal of this package certainly does not mean it is not valuable– for example, the Excel-based DIA has been used to inform Federal Energy Regulatory Committee licensing activities at hydroelectric dams on the Penobscot. Having a powerful yet accessible open-source tool to support the transparency and reproducibility of the science behind federal management actions would be valuable even if it can only be used in specific scenarios.

**General checks**

- _Substantial scholarly effort:_ As discussed in the paragraph above, this submission meets the scope eligibility described in the JOSS guidelines. However, I think it would be prudent to make it slightly more obvious that the project is currently only applicable to Penobscot River Atlantic salmon. For example, the Statement of Need section in the paper states “We created dia for use by fisheries researchers, managers, and practitioners interested in understanding population dynamics of intensively managed endangered Atlantic salmon in the USA.” The phrase “in the USA” is likely an overgeneralization of the package’s current functionality.

- _Reproducibility:_ The figures in the paper could not be exactly reproduced without knowledge of the seed used for the random number generation. Similarly, I would recommend including `set.seed()` at the beginning of the README and help file examples to emphasize to users that otherwise results would not be reproducible. Also, when running the second README example I had to change it slightly in order to reproduce it on my computer: my laptop only has 8 cores, so I changed `ncpus <- 10` to `ncpus <- parallel::detectCores()-1. `

**Documentation**
- _A statement of need:_ The Overview section of the README could be expanded to be more explicit about “what problems the software is designed to solve and who the target audience is.”
- _Functionality documentation:_ While the core functionality of the software is documented to a satisfactory level, I wanted to note that as someone who mainly uses future/furrr for parallelization, I had to reference the snowfall documentation to fully understand the parallel execution example in the README.
- _Community guidelines:_ The package currently lacks “clear guidelines for third parties wishing to 1. contribute to the software 2. report issues or problems with the software 3. seek support.” Please provide these instructions in the README or within a dedicated CONTRIBUTING.md file linked in the README.

**Software paper**
- _State of the field:_ It might be valuable to relate the functionality of this package to other software tools used for similar applications, especially the shadia package developed by the submitting author. In particular, emphasizing differences between the two packages besides the focal species could be useful for readers to quickly understand why DIA was developed as a separate project.
- _Quality of writing:_ Please remove ‘Figure1’ from the image caption in the paper.md file, as “Figure 1” is automatically added before the caption when the file is rendered to a PDF. (Also note minor typo in the README: shownd instead of shown.)
- _References:_
     - Please ensure that in-text citations are correctly formatted. For example, “(e.g., @noaa:2013; @nieland:2020)” renders as (e.g., National Marine Fisheries Service (2013); Nieland & Sheehan (2020)). Changing the markdown syntax  to “[e.g., @noaa:2013; @nieland:2020]” will remove the nested parentheses and hyperlink the entire citation instead of just the year, ensuring consistency with the in-text citations in the previous sentences.
     - Please update the year in the R Core Team citation from 2022 to 2024.
     - Please add a list of references to the README for the cited works (“Nieland et al. 2013, 2015; Nieland and Sheehan 2020”) and/or add hyperlinks to the relevant documents.

**Other comments**
I was wondering when it would be necessary to have `stillwater_use_old` included in the package?
