### 2024-11-13 dd4ec36
Update paper.bib
Mainly adjustments to capitalization
