+++ paper.bib
@@ -11,8 +11,8 @@
 @misc{noaa:2013,
   	url = {https://repository.library.noaa.gov/view/noaa/55737},
   	Author = {{National Marine Fisheries Service}},
-	  Publisher = {National Oceanic and Atmospheric Administration},
-  	Title = {National Marine Fisheries Service Endangered Species Act Biological Opinion Amendment of License for the Mattaceunk Project F/NER/2013/9640},
+	  Publisher = {{National Oceanic and Atmospheric Administration}},
+  	Title = {{National Marine Fisheries Service} {Endangered Species Act} {Biological Opinion} {Amendment of License} for the {Mattaceunk Project} F/NER/2013/9640},
   	Year = {2013}
 }
 
@@ -20,7 +20,7 @@
   	url = {https://repository.library.noaa.gov/view/noaa/23714},
   	Author = {{Nieland}, J.~L. and {Sheehan}, T.~F.},
   	Journal = {Northeast Fisheries Science Center Reference Document 19-16},
-	  Publisher = {National Oceanic and Atmospheric Administration},
+	  Publisher = {{National Oceanic and Atmospheric Administration}},
   	Title = {Quantifying the effects of dams on {Atlantic} salmon in the {Penobscot River} watershed, with a focus on {Weldon Dam}},
   	Year = {2020},
   	doi = {10.25923/v67x-kk62}
@@ -122,7 +122,7 @@
 }  
   
 @Book{ggplot2,
-  author = {Hadley Wickham},
+  author = {Wickham, H.},
   title = {{ggplot2}: Elegant graphics for data analysis},
   publisher = {Springer-Verlag New York},
   year = {2016},
