### 2024-11-14 3460b31
Update README.md from rmk118 JOSS Review
Update README to include changes suggested by rmk118 at [openjournals/joss-reviews/issues/7475](https://github.com/openjournals/joss-reviews/issues/7475):

Expanded overview to include context, address typographical errors, additional commenting of `snowfall` parallel workflow for unfamiliar users, set random seeds (7475) for reproducible examples, and add list of references and hyperlinks for references in README.md
