+++ README.md
@@ -4,7 +4,7 @@
 
 ## Overview
 
-An R package for hosting code and data, and running the Atlantic salmon Dam Impact Analysis (DIA) model.
+An R package for hosting code and data, and running the Atlantic salmon Dam Impact Analysis (DIA) model. This package was designed to provide a transparent, open-source version of the National Oceanic and Atmospheric Administration's Dam Impact Analysis (DIA) model for Atlantic salmon (*Salmo salar*) in the Penobscot River, Maine, USA. We created the package for use by fisheries scientists, managers, and decision makers as a transparent tool to support decision-making related to conservation of Atlantic salmon in this priority waterbody. It uses life-history data, flow-specific survival rates and other empirical estimates to simulate the response of Atlantic salmon in the Penobscot River over multiple generations to key conservation and restoration decisions such as supplemental fish stocking and passage through hydropower dams.
 
 ## Installation
 
@@ -27,21 +27,18 @@ To install `dia`, you will need to have `devtools` installed ahead of time in R,
 `tests/` Includes package tests for default parameter accuracy conducted on package build
 
 ## Getting started
-There are two primary user-facing functions within the `dia` package. The 
-`run_dia()` function can be called within an Rscript to run the Dam Impact Analysis (DIA) model using default or user-defined inputs. The `run_dia_shiny()` function launches a graphical user interface (GUI) that opens in a web browser and is written with the `shiny` package. The former provides functionality for large-scale simulation studies in R whereas the latter provides an intuitive  interface for running smaller numbers of management scenarios.
+There are two primary user-facing functions within the `dia` package. The `run_dia()` function can be called within an Rscript to run the Dam Impact Analysis (DIA) model using default or user-defined inputs. The `run_dia_shiny()` function launches a graphical user interface (GUI) that opens in a web browser and is written with the `shiny` package. The former provides functionality for large-scale simulation studies in R whereas the latter provides an intuitive  interface for running smaller numbers of management scenarios.
 
 ### `run_dia()`
-The `run_dia()` function provides the primary user-facing function for the `dia` package. The purpose of the function is to run the Dam Impact Analysis for some number of generations (15 by default). The output of this function is a  dataframe (ie. data table) containing number of wild and hatchery-origin adult returns in each generation. All default argument values (model inputs) are  based on the most-recent version of the NOAA Dam Impact Analysis Excel-based model (v67, Nieland et al. 2013, 2015; Nieland and Sheehan 2020).
+The `run_dia()` function provides the primary user-facing function for the `dia` package. The purpose of the function is to run the Dam Impact Analysis for some number of generations (15 by default). The output of this function is a  dataframe (i.e., data table) containing number of wild and hatchery-origin adult returns in each generation. All default argument values (model inputs) are  based on the most-recent version of the NOAA Dam Impact Analysis Excel-based model (v67, [Nieland et al. 2013](https://repository.library.noaa.gov/view/noaa/55737); [Nieland et al. 2015](https://academic.oup.com/icesjms/article-pdf/72/8/2423/31226690/fsv083.pdf); [Nieland and Sheehan 2020](https://repository.library.noaa.gov/view/noaa/23714)).
 
[…]
