### 2024-11-14 2a45b08
Update README.md
correct "and intuitive user interface"
