### 2024-11-14 9cab13e
Update README.md
Couple missing hyperlinks for references in examples
