### 2024-11-14 2b8d169
Update citation for R: openjournals/joss-reviews#7475
