+++ paper.bib
@@ -117,7 +117,7 @@
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2022},
+  year = {2024},
   url = {https://www.R-project.org/}
 }  
   
