### 2024-11-14 e013903
update Figure1 @openjournals/joss-reviews#7475
