### 2024-11-14 a7b0e57
add Contributing.md openjournals/joss-reviews#7475
