+++ Contributing.md
@@ -1,0 +1,1 @@
+We encourage contributions from the community and welcome collaborations for extension to other priority conservation waterbodies for Atlantic salmon. You can ask questions or submit bug reports through the [issues page](https://github.com/danStich/dia/issues). Please search issues and ensure that a similar issue does not already exist before submitting a new issue. If you would like to contribute improvements or modifications directly to the repository, you can fork the repository and submit a merge request so those contributions can be reviewed.
