### 2024-11-14 4117d68
update `stillwater_use_old` description openjournals/joss-reviews#7475
