+++ R/data.R
@@ -207,7 +207,11 @@
 #' @description Stillwater use table from the Atlantic salmon 
 #' Dam Impact analysis model v67. Previous values used for 
 #' Cumulative probabilities of proportional use of Stillwater Branch by 
-#' outmigrating Atlantic salmon smolts in the Penobscot River, ME.
+#' outmigrating Atlantic salmon smolts in the Penobscot River, ME. This 
+#' version can be used to reproduce analyses from Nieland et al. (2013, 2015)
+#' and to explore current scenarios relative to historical conditions (e.g., 
+#' conditions prior to removal of dams and implementation of new fish
+#' passage systems).
 #' 
 #' @format A dataframe with 5,000 observations of 2 variables
 #' \describe{
