### 2025-01-06 36f1854
Update Contributing.md
Improved Contributing section
