### 2025-01-06 d689a3e
Update Contributing.md
Resolve issue #3
