### 2025-01-06 #7475 @danStich
Thanks very much! I have meetings much of today but will begin working to address these changes.

Dan

----
Daniel S. Stich (he/him)
Associate Professor, Biology Department and
Secretary-Treasurer, NY Chapter American Fisheries Society
113A Perna Science, SUNY Oneonta, NY 13820
Office: 607-436-3734
Cell: 518-860-4107
Email: ***@***.***
Website: danstich.github.io/stich<https://danstich.github.io/stich/index.html>
