### 2025-01-14 083e25b
update documentation formatting for datasets and package-level docs for pkgdown site openjournals/joss-reviews#7475
