+++ R/dia.R
@@ -12,60 +12,57 @@
 #' The following functions can be called directly from dia.
 #' 
 #'  \describe{
-#'    \code{\link{run_dia}}{ The primary user-facing function used to run the multi-generation DIA model for Atlantic salmon} \cr \cr
-#'    \code{\link{run_dia_shiny}}{ A `shiny` implementation of `run_dia()` with graphical user interface} \cr \cr
-#'    \code{\link{run_one_gen}}{ Run the DIA model for a single generation} \cr \cr
-#'    \code{\link{get_dam_passage}}{ Get flow-correlated downstream dam passage values and probability of using Stillwater Branch for smolts each year in `run_one_gen`} \cr \cr
-#'    \code{\link{make_egg_to_smolt_survival}}{ Simulate survival from egg to smolt stage for naturally produced smolts in `run_one_gen`} \cr \cr
-#'    \code{\link{make_eggs_per_female}}{ Simulate number of eggs per female each year in `run_one_gen`} \cr \cr
-#'    \code{\link{make_inriver_s}}{ Simulate natural, in-river survival per km for downstream migrating smolts in `run_one_gen`} \cr \cr   
-#'    \code{\link{make_marine_s}}{ Simulate marine survival each year in `run_one_gen`} \cr \cr    
-#'    \code{\link{run_downstream_passage}}{ Run the downstream smolt passage module in `run_one_gen`} \cr \cr
-#'    \code{\link{run_inefficiency}}{ Re-distribute adult upstream migrants following failed dam passage in `run_one_gen`} \cr \cr    
-#'    \code{\link{run_straying}}{ Assign destination production units for adult returns in `run_one_gen`} \cr \cr      
-#'    \code{\link{run_upstream_passage}}{ Function used to run the upstream passage module for adults in `run_one_gen`} \cr \cr        
-#'    \code{\link{smolt_stocking}}{ Function used to distribute stocked smolts among production units in `run_one_gen`} \cr \cr
+#'    \item{\code{\link{run_dia}}}{The primary user-facing function used to run the multi-generation DIA model for Atlantic salmon}
+#'    \item{\code{\link{run_dia_shiny}}}{A `shiny` implementation of `run_dia()` with graphical user interface}
+#'    \item{\code{\link{run_one_gen}}}{Run the DIA model for a single generation}
+#'    \item{\code{\link{get_dam_passage}}}{Get flow-correlated downstream dam passage values and probability of using Stillwater Branch for smolts each year in `run_one_gen`}
+#'    \item{\code{\link{make_egg_to_smolt_survival}}}{Simulate survival from egg to smolt stage for naturally produced smolts in `run_one_gen`}
+#'    \item{\code{\link{make_eggs_per_female}}}{Simulate number of eggs per female each year in `run_one_gen`}
+#'    \item{\code{\link{make_inriver_s}}}{Simulate natural, in-river survival per km for downstream migrating smolts in `run_one_gen`}   
+#'    \item{\code{\link{make_marine_s}}}{Simulate marine survival each year in `run_one_gen`}    
+#'    \item{\code{\link{run_downstream_passage}}}{Run the downstream smolt passage module in `run_one_gen`}
+#'    \item{\code{\link{run_inefficiency}}}{Re-distribute adult upstream migrants following failed dam passage in `run_one_gen`}    
+#'    \item{\code{\link{run_straying}}}{Assign destination production units for adult returns in `run_one_gen`}      
+#'    \item{\code{\link{run_upstream_passage}}}{Function used to run the upstream passage module for adults in `run_one_gen`}        
[…]
