### 2025-01-17 #7475 @danStich
Hi @cheginit, I am still working to address a couple of @Fabbiologia 's concerns. I circulated a draft conceptual diagram with coauthors and have wrote two new applied examples for README.md that I have not yet pushed. And, I am working on editing the text to address concerns about scope.
