### 2025-01-20 868555e
update documentation and meta data for all built-in datasets for openjournals/joss-reviews#7475
