+++ R/data.R
@@ -41,13 +41,13 @@
 #' @title Straying locations
 #' 
 #' @description Straying matrix from the Atlantic salmon Dam 
-#' Impact Analysis model v67. Straying locations of adult Atlantic 
-#' salmon based on behavioral patterns. 
+#' Impact Analysis model v67. Straying locations (PUs) of returning adult Atlantic 
+#' salmon based on natal PU. 
 #' 
 #' @format A dataframe with 15 observations of 16 variables.
 #' \describe{
 #'   \item{\code{Natal_PU}}{Natal production unit as smolt}
-#'   \item{\code{PU_01, ..., PU14}}{Destination production unit as adult}
+#'   \item{\code{PU_01, ..., PU14}}{Probability of returning to each destination production unit as an adult conditioned on natal PU}
 #' }
 #' 
 #' @source 
@@ -69,9 +69,14 @@
 
 #' @title Straying matrix
 #' 
-#' @description Numeric columns only from \code{\link{straying_locations}}
+#' @description Numeric columns only from \code{\link{straying_locations}}. A 
+#' matrix of return probabilities to destination PUs (columns) conditioned on 
+#' natal PU (rows) where the diagonal represents probability of returning to 
+#' natal PU (i.e., homing rate). These values are used in combination with the
+#' \code{\link{inefficiency_matrix}} to spatially distribute returning adult 
+#' spawners in each generation of simulations.
 #' 
-#' @format A matrix with 15 observations of 16 variables.
+#' @format A matrix with 15 observations of 15 variables.
 #' 
 #' @source 
 #' Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
@@ -95,15 +100,18 @@
 #' @description Upstream inefficiency table from the Atlantic salmon 
 #' Dam Impact Analysis model v67. When adult Atlantic salmon do not 
 #' pass a dam, a proportion of those fish die, a proportion return 
-#' to sea, and a proportion go elsewhere to attempt to spawn.
+#' to sea, and a proportion go elsewhere to attempt to spawn. These values are
+#' used in combination with the \code{\link{straying_matrix}} to spatially
+#' distribute returning adult spawners in each generation of simulations.
+#' 
 #' 
 #' @format A dataframe with 15 observations of 15 variables
 #' \describe{
 #'   \item{\code{dam_failed_to_pass}}{The dam the adults fail to pass}
 #'   \item{\code{p_die}}{The proportion of failed migrants that die}
 #'   \item{\code{p_return_to_sea}}{The proportion of failed migrants that return to sea}
-#'   \item{\code{p_go_elsewhere}}{The proportion going elsewhere}
-#'   \item{\code{PU_01...PU14}}{The destination PUs for failed migrants at each dam with proportion going to each}
+#'   \item{\code{p_go_elsewhere}}{The proportion of failed migrants going elsewhere}
+#'   \item{\code{PU_01...PU14}}{The destination PUs for failed migrants at each dam with proportion of failed migrants going to each}
 #' }
 #' 
 #' @source 
@@ -127,7 +135,9 @@
 #' 
 #' @description Numeric columns 5 - 20 from \code{\link{upstream_inefficiency}}
 #' 
-#' @format A matrix with 15 observations of 15 variables.
+#' @format A matrix with 15 observations of 15 variables. 
+#' 
+#' @details See \code{\link{upstream_inefficiency}} for additional details.
 #' 
 #' @source
 #' Nieland JL, Sheehan TF. 2020. Quantifying the Effects of Dams on Atlantic Salmon
@@ -151,7 +161,9 @@
 #' @description Life Stage Survival table from the Atlantic salmon 
[…]
