### 2025-01-20 f577a35
update run_dia function documentation for openjournals/joss-reviews#7475
