+++ man/run_dia.Rd
@@ -31,56 +31,64 @@ run_dia(
 )
 }
 \arguments{
-\item{n_generations}{Number of generations to simulate for simulation. One 
+\item{n_generations}{Number of generations to simulate. One 
 complete generation from egg to returning adult is six years.}
 
-\item{n_wild}{Number of starting wild adult salmon.}
+\item{n_wild}{Number of starting wild, adult Atlantic salmon.}
 
-\item{n_hatchery}{Number of starting hatchery adult salmon.}
+\item{n_hatchery}{Number of starting hatchery adult Atlantic salmon.}
 
 \item{stocking}{Numeric indicating whether hatchery stocking is on (`1`) or
-off (`0`).}
+off (`0`). If hatchery stocking is on, then hatchery smolts are stocked into
+production units based on the \code{proportion_stocked} column in the
+built-in \code{\link{stocking_default}} dataset.}
 
 \item{n_stocked}{A numeric vector corresponding to `n_generations` in length
-with number of stocked smolts per generation.}
+with the number of stocked hatchery smolts per generation.}
 
 \item{upstream}{Upstream passage efficiency of adult Atlantic salmon through
 dams. Default values are based on NMFS (2013) prescriptions as implemented
-in Nieland and Sheehan (2020).}
+in Nieland and Sheehan (2020). Any numeric value between zero and 1, 
+inclusive, can be specified for any dam by the user.}
 
 \item{downstream}{Downstream survival of Atlantic salmon smolts through dams.
 The default values are based on the most recent values used by Nieland et al. 
 (2020) based on standards established in the the NOAA Species Protection Plan
 following implementation of the Penobscot River Restoration Project.
-The default `NULL` randomly samples correlated survival rates of smolts for 
+The default `NA` randomly samples correlated survival rates of smolts for 
 each dam based on cumulative flow probabilities and associated empirical 
-survival rates (Nieland et al. 2013, Nieland and Sheehan 2020).}
+survival rates (Nieland et al. 2013, Nieland and Sheehan 2020), and can also
+be used for dams that have a default value other than \code{NA}.}
 
 \item{in_river_s}{In-river survival per kilometer for downstream migrating
-smolts. The default `NA` value simulates from cumulative distribution 
-function using values in \code{\link{in_river_m}}.}
+smolts. The default `NA` value simulates from a cumulative distribution 
+function using values in \code{\link{in_river_m}}. This value is the same
+for hatchery stocked and wild smolts.}
 
 \item{mattaceunk_impoundment_mortality}{Mortality incurred by Atlantic salmon
 smolts during migration through the Mattaceunk (Weldon) Dam impoundment. The
-default value is based on Nieland and Sheehan (2020). Based on results of 
+default value is based on Nieland and Sheehan (2020), who used results of 
 Holbrook et al. (2011) and Stich et al. (2015a).}
 
 \item{p_stillwater}{Probability that fish use the Stillwater Branch for 
-downstream migration. The default (`NULL`) draws flow-correlated values from
-a cumulative distribution of flows with paired estimates of p_stillwater
-used by Nieland and Sheehan (2020), based on empirical results in Holbrook et al.
-(2006), Stich et al. (2014), and Stich et al. (2015a).}
+downstream migration. The default (\code{NULL}) draws flow-correlated probability
+of using the Stillwater Branch from a cumulative distribution of flows with 
+paired estimates of \code{p_stillwater} used by Nieland and Sheehan (2020), 
+based on empirical results in Holbrook et al. (2006), Stich et al. (2014), 
+and Stich et al. (2015a).}
 
 \item{indirect_latent_mortality}{Indirect, latent mortality incurred by 
[…]
