### 2025-01-20 5fb41f0
Update readme to include additional background, workflow diagram, and additional examples for openjournals/joss-reviews#7475
