### 2025-01-20 713f625
increment RoxygenNote for openjournals/joss-reviews#7475
