+++ DESCRIPTION
@@ -12,7 +12,7 @@ License: GPL-3 + file LICENSE
 Encoding: UTF-8
 Imports: truncnorm, tidyr, shiny
 LazyData: true
-RoxygenNote: 7.2.1
+RoxygenNote: 7.3.1
 Suggests: 
     testthat (>= 3.0.0), DT, dplyr, ggplot2
 Config/testthat/edition: 3
