### 2025-01-20 98b69fe
update paper and bibliography files for openjournals/joss-reviews#7475
