+++ paper.md
@@ -30,19 +30,19 @@ bibliography: paper.bib
 
 # Summary
 
-Populations of anadromous (sea-run) fishes such as Atlantic salmon *Salmo salar* have experienced severe global declines as a result of pollution, overfishing, and construction of dams [@limburg:2009]. Life-history-based simulation models are commonly used for planning and implementing fisheries recovery activities for many diadromous species [e.g., @nieland:2013; @barber:2018; @stich:2019; @zydlewski:2021]. Prior to the expansion of open-source development, many institutionalized decision-support tools utilizing such models relied on closed-source or paid software. For example, the National Oceanic and Atmospheric Administration (NOAA) Dam Impact Analysis (DIA) was originally created as a stochastic life cycle model for Atlantic salmon in the `@RISK` add-in within Microsoft Excel [@nieland:2013]. This tool differs from those hosted in similar R packages for non-salmonid species [e.g. @stich:2019; @zydlewski:2021] in that it employs homing and straying probabilities to natal streams for adult spawning and integrates results of physical modeling to inform population dynamics [@nieland:2013; @nieland:2020]. We created the `dia` package [@dia] for the R programming language [@R] as a freely accessible, open-source implementation of these tools that will promote transparency in planning and decision making and improve the availability and quality of these tools over time.
+Populations of anadromous (sea-run) fishes such as Atlantic salmon *Salmo salar* have experienced severe global declines as a result of pollution, overfishing, and construction of dams [@limburg:2009]. Life-history-based simulation models are commonly used for planning and implementing fisheries recovery activities for many diadromous species [e.g., @nieland:2013; @barber:2018; @stich:2019; @zydlewski:2021]. Prior to the expansion of open-source development, many institutionalized decision-support tools utilizing such models relied on closed-source or paid software. For example, the National Oceanic and Atmospheric Administration (NOAA) Dam Impact Analysis (DIA) was originally created as a stochastic life cycle model for Atlantic salmon in the `@RISK` add-in within Microsoft Excel [@nieland:2013]. This tool differs from those hosted in similar R packages for non-salmonid species [e.g. @stich:2019; @zydlewski:2021] in that it employs homing and straying probabilities to natal streams for adult spawning and integrates results of physical modeling to inform population dynamics [@nieland:2013; @nieland:2020]. This class of tools, in general, provides advantages for decision making related to anadromous species because it allows integration of geographically and temporally explicit stock dynamics (e.g., influences of dams) that are not readily implemented in classical fisheries stock assessment tools such as those available in existing R packages [e.g. @FLR; @Erickson2022; @FSA]. We created the `dia` package [@dia] for the R programming language [@R] as a freely accessible, open-source implementation of these tools that will promote transparency in planning and decision making and improve the availability and quality of these tools over time.
 
 # Statement of need
 
-`dia` is an R-based implementation of a previously closed-source life cycle model of Atlantic salmon population dynamics that is used to understand the sensitivity of species recovery trajectories to natural and anthropogenic marine and freshwater influences. It was developed to assess the sensitivity of restoration outcomes to uncertainty in life-history inputs alongside the impacts of dams and restoration activities. The DIA model uses empirical life-history estimates (e.g., survival), predictive flow and resulting flow-specific dam survival modeling, and other empirical data in freshwater and marine environments to simulate consecutive generations of Atlantic salmon in the Penobscot River (Maine, USA) under varying environmental conditions or management decisions. Management decisions include fish passage rates at dams, fishery harvest rates, and numbers and locations for hatchery stocking of fish [@nieland:2013]. Since development, it has been used for mechanistic exploration of key life-history uncertainties within the context of species recovery [@nieland:2015] and to support decision-making at federally regulated hydropower dams on the Penobscot River [e.g., @noaa:2013; @nieland:2020].
[…]
