+++ paper.bib
@@ -140,4 +140,32 @@
     number = {43},
     pages = {1686},
     doi = {10.21105/joss.01686},
-  }
+}
+  
+@Article{FLR,
+    title = {FLR: an open-source framework for the evaluation and development of management strategies},
+    author = {L.T. Kell and I. Mosqueira and P. Grosjean and J-M. Fromentin and D. Garcia and R. Hillary and E. Jardim and S. Mardle and M.A. Pastoors and J.J. Poos and F. Scott and R.D. Scott},
+    journal = {ICES Journal of Marine Science},
+    year = {2007},
+    volume = {64},
+    pages = {640-646}
+}  
+
+@Manual{FSA,
+    title = {FSA: Fisheries Stock Analysis},
+    author = {Derek H. Ogle and Jason C. Doll and Powell Wheeler and Alexis Dinno},
+    year = {2022},
+    note = {R package version 0.9.3},
+    url = {https://github.com/fishR-Core-Team/FSA}
+}  
+  
+  
+@article{Erickson2022, 
+  doi = {10.21105/joss.03444}, 
+  url = {https://doi.org/10.21105/joss.03444}, 
+  year = {2022}, publisher = {The Open Journal}, 
+  volume = {7}, number = {71}, 
+  pages = {3444}, 
+  author = {Richard A. Erickson and Daniel S. Stich and Jillian L. Hebert}, 
+  title = {fishStan: Hierarchical Bayesian models for fisheries}, 
+  journal = {Journal of Open Source Software} }  
