### 2025-01-20 #7475 @danStich
Hi @Fabbiologia,

I think I've finished pushing changes to the manuscript, README, and R package documentation in response to your suggestions. I also wrote responses to each of your queries as comments in the [issue tracker[(https://github.com/danStich/dia/issues).

I did not close any of the issues because I wanted to give you a chance to look through the changes. Please do let me know if you'd rather I'd close those issues for now, or if you'd rather close them once you've had a chance to review the changes and comments.
