### 2025-01-20 1f881d0
Add title and license badge to README for openjournals/joss-reviews#7475 and confirmed repository license matches software license
