### 2025-01-20 d30d9e0
increment version number for JOSS Zenodo release
