+++ README.md
@@ -12,7 +12,7 @@ An R package for the National Oceanic and Atmospheric Administration dam impact 
 - [Examples](#examples)
 - [References](#references)
 
-# Overview
+## Overview
 An R package for hosting code and data, and running the Atlantic salmon Dam Impact Analysis (DIA) model. This package was designed to provide a transparent, open-source version of the National Oceanic and Atmospheric Administration's Dam Impact Analysis (DIA) model for Atlantic salmon (*Salmo salar*) in the Penobscot River, Maine, USA, that was previously hosted as an @RISK add on for Microsoft Excel. We created the package for use by fisheries scientists, managers, and decision makers as a transparent tool to support decision-making related to conservation of Atlantic salmon in this priority waterbody that hosts the majority of adult returns to the USA each year. It uses life-history data, flow-specific survival rates and other empirical estimates to simulate the response of Atlantic salmon in the Penobscot River over multiple generations to key conservation and restoration decisions such as supplemental fish stocking and passage through hydropower dams. A number of example uses are provided below that build in complexity.
 
 The full details of the model parameterization, including background information about the Atlantic salmon life cycle and the Penobscot River, are available from [Nieland et al. (2013)](https://repository.library.noaa.gov/view/noaa/4559), with a specific case study by [Nieland et al. (2015)](https://academic.oup.com/icesjms/article/72/8/2423/2458708) and updated biological data and a second case study in [Nieland and Sheehan (2020)](https://repository.library.noaa.gov/view/noaa/23714). Below is a comparison of the original model parameterized by Nieland et al. (2013) and the core workflow of the `dia` package. Helper functions and built-in data sets are used to simulate a population through space in a single generation using the `run_one_gen()` function, which is then wrapped in the `run_dia()` function that facilitates stochastic simulation over a user-specified number of generations.
