### 2025-01-20 #7475 @danStich
I've incremented the version number, published the release and set as latest in GitHub repository, and archived the release on Zenodo.

Updated software version number: v1.0.1
DOI from release archive on Zenodo: 10.5281/zenodo.14708353
