### 2025-01-20 #7475 @cheginit
For the paper, please address the following:

- [ ] **L12-13:** "Prior to the expansion of open-source development" is verbose. Consider simplifying to "Before open-source tools became prevalent"
- [ ] **L18-19:** Clarify "homing and straying probabilities" by providing a brief definition or context.
- [ ] **L27:** "improve the availability and quality of these tools" is vague. Specify how the quality is improved. Do you mean `dia` will improve other tools?
- [ ] **L32:** The phrase "understand the sensitivity of species recovery trajectories to natural and anthropogenic marine and freshwater influences" could be simplified for clarity.
- [ ] **L51:** Avoid repetition of "allows exploration" in "allows exploration of future restoration scenarios through... and allows exploration of uncertainty."
- [ ] **L53-54:** Break down the description of the `run_dia()` function for better readability.
- [ ] **L59-60:** Combine sentences about `run_dia_shiny()` to avoid redundancy.
- [ ] **L60:** The citation is not correct "(Chang et al. (2022))".
- [ ] **L64:** "Any other R or shiny application" could be replaced with "other R or Shiny applications."
- [ ] **L78:** For the NOAA fund if you are needed to provide a grant number do so, if not, leave as-is.

- [ ] **L78:** Add the missing DOI if that's correct `10.1093/icesjms/fsm012`
- [ ] **L78:** Prefer using DOI instead of repos. For this item, the developers already provided a Zenodo, use it: https://zenodo.org/records/6098468
