### 2025-01-20 a43861c
update paper, references, and author information in response to editor comments for openjournals/joss-reviews#7475
