+++ paper.bib
@@ -104,10 +104,10 @@
 }
 
 @Manual{shiny,
-  title = {shiny: Web application framework for {R}},
-  author = {Chang, W. and Cheng, J. and Allaire, J.~J. and Sievert, C. and Schloerke, B. and Xie, Y. and Allen, J. and McPherson, J. and Dipert, A. and Borges, B.},
-  year = {2022},
-  note = {R package version 1.7.3},
+  title = {shiny: Web Application Framework for R},
+  author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+  year = {2024},
+  note = {R package version 1.10.0},
   url = {https://CRAN.R-project.org/package=shiny},
   doi = {10.32614/CRAN.package.shiny}
 }
@@ -148,7 +148,8 @@
     journal = {ICES Journal of Marine Science},
     year = {2007},
     volume = {64},
-    pages = {640-646}
+    pages = {640-646},
+    doi = {10.1093/icesjms/fsm012}
 }  
 
 @Manual{FSA,
@@ -156,7 +157,8 @@
     author = {Derek H. Ogle and Jason C. Doll and Powell Wheeler and Alexis Dinno},
     year = {2022},
     note = {R package version 0.9.3},
-    url = {https://github.com/fishR-Core-Team/FSA}
+    url = {https://github.com/fishR-Core-Team/FSA},
+    doi = {10.5281/zenodo.597719}
 }  
   
   
