### 2025-01-21 #7475 @danStich
@cheginit I've made the corrections and changes that you suggested. Line-specific responses are provided below for the paper, references, and author information (ORCIDs).

**Paper:**

L12-13: "Prior to the expansion of open-source development" is verbose. Consider simplifying to "Before open-source tools became prevalent"

* Changed as suggested.

L18-19: Clarify "homing and straying probabilities" by providing a brief definition or context.

* Changed this sentence to: "This tool differs from those hosted in similar R packages for non-salmonid species in that it incorporates homing behavior (probability of adults returning to natal streams or straying to others) and it integrates results of physical modeling to inform population dynamics."

L27: "improve the availability and quality of these tools" is vague. Specify how the quality is improved. Do you mean dia will improve other tools?

* I think we were "hinting" at potential for future development. But, we updated the manuscript to address this directly in response to one of the reviewers, so we've removed it from this sentence.

L32: The phrase "understand the sensitivity of species recovery trajectories to natural and anthropogenic marine and freshwater influences" could be simplified for clarity.

* Simplified as: "understand population response to natural and anthropogenic influences in freshwater and marine environments" to improve clarity.

L51: Avoid repetition of "allows exploration" in "allows exploration of future restoration scenarios through... and allows exploration of uncertainty."

* Rephrased to avoid repetition and be more concise: "However, it also allows exploration uncertainty associated with life-history parameters and investigation of future restoration scenarios through a variety of user-facing options."

L53-54: Break down the description of the run_dia() function for better readability.

* Split this sentence into 3 shorter sentences.

L59-60: Combine sentences about run_dia_shiny() to avoid redundancy.

* Combined these two sentences into one.

L60: The citation is not correct "(Chang et al. (2022))".

* We have updated with the citation for the latest version (Change et al. 2024) and replaced parentheses with brackets so the in-text citation renders correctly.

L64: "Any other R or shiny application" could be replaced with "other R or Shiny applications."

* Change made.

L78: For the NOAA fund if you are needed to provide a grant number do so, if not, leave as-is.

* Added the sponsor award number in parentheses.

**References:**

L78: Add the missing DOI if that's correct 10.1093/icesjms/fsm012

* Added DOI to the bibtex citation.

L78: Prefer using DOI instead of repos. For this item, the developers already provided a Zenodo, use it: https://zenodo.org/records/6098468

* Added DOI from Zenodo

**Author Information**
* Added ORCID for T Sheehan, J Nieland does not currently have an ORCID. I also
added the ORCID for T Sheehan to the DESCRIPTION file for the R package.
