### 2025-01-21 #7475 @danStich
Thank you for noticing this. I changed the metadata in the Zenodo archive so the title matches the JOSS submission title and the correct creators are listed with ORCIDS for those who have them.

----
Daniel S. Stich (he/him)
Associate Professor, Biology Department and
Secretary-Treasurer, NY Chapter American Fisheries Society
113A Perna Science, SUNY Oneonta, NY 13820
Office: 607-436-3734
Cell: 518-860-4107
Email: ***@***.***
Website: danstich.github.io/stich<https://danstich.github.io/stich/index.html>
