### 2025-01-21 #7475 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
