### 2025-01-21 336e878
correct typos for openjournals/joss-reviews#7475: missing "of" in second paragraph and extra "the" in third paragraph of "statement of need"
