+++ paper.md
@@ -37,9 +37,9 @@ Populations of anadromous (sea-run) fishes such as Atlantic salmon *Salmo salar*
 
 `dia` is an R-based implementation of a previously closed-source life cycle model of Atlantic salmon population dynamics that is used to understand population response to natural and anthropogenic influences in freshwater and marine environments. It was developed to assess the sensitivity of restoration outcomes to uncertainty in life-history inputs alongside the impacts of dams and restoration activities. The DIA model uses empirical life-history estimates (e.g., survival), predictive flow and resulting flow-specific dam survival modeling, and other empirical data in freshwater and marine environments to simulate consecutive generations of Atlantic salmon in the Penobscot River (Maine, USA) under varying environmental conditions or management decisions. As the largest remaining population of this critically endangered species in the USA, the population is intensively managed. Management decisions include fish passage rates at dams, fishery harvest rates, and numbers and locations for hatchery stocking of fish [@nieland:2013]. Since development, it has been used for mechanistic exploration of key life-history uncertainties within the context of species recovery [@nieland:2015] and to support decision-making at federally regulated hydropower dams on the Penobscot River [e.g., @noaa:2013; @nieland:2020].
 
-We created `dia` for use by fisheries researchers, managers, and practitioners interested in understanding population dynamics of intensively managed endangered Atlantic salmon in the USA. The R package maintains the core routines from the original closed-source version of the model by replicating spreadsheet-based calculations, and incorporates original data and parameter sets as built-in objects that serve as default values for arguments of the primary user-facing functions. However, it also allows exploration uncertainty associated with life-history parameters and investigation of future restoration scenarios through a variety of user-facing options.
+We created `dia` for use by fisheries researchers, managers, and practitioners interested in understanding population dynamics of intensively managed endangered Atlantic salmon in the USA. The R package maintains the core routines from the original closed-source version of the model by replicating spreadsheet-based calculations, and incorporates original data and parameter sets as built-in objects that serve as default values for arguments of the primary user-facing functions. However, it also allows exploration of uncertainty associated with life-history parameters and investigation of future restoration scenarios through a variety of user-facing options.
 
-The two primary user-facing functions within the `dia` package are the `run_dia()` and `run_dia_shiny()`, which provide redundant interfaces for using Dam Impact Analysis (DIA) models in different ways. The `run_dia()` function provides an extensible interface to DIA. It can be used for long-run simulation or decision-optimization studies. It allows incorporation of user-specified data sets such as flow-correlated survival probabilities at dams and in free-flowing river reaches, marine survival and other life-history inputs, or fish-stocking data. The `run_dia_shiny()` function deploys a graphical user interface using the `shiny` package [@shiny] that is less extensible but more easily used by fishery managers and practitioners who may be less familiar with programming and it also includes exportable results from simulation models including `.csv` or other flat-file formats and default plots through the `ggplot2` R package [@ggplot2; @tidyverse; Figure 1]. Both can be deployed on networked servers as other R or `shiny` applications to improve accessibility or facilitate use on high performance computers for large simulations. The GitHub repository [@dia] includes additional instructions for installation and a variety of potential uses of `run_dia()` and `run_dia_shiny()`, with shorter examples included in the package help files. While implementation is currently limited to the Penobscot River as a priority conservation water in the USA, the package serves as one example to help generalize these modeling approaches to Atlantic salmon and other sea-run fish in watersheds globally. Specifically, while many of the built-in datasets and helper functions in `dia` are generalized or could be used to simulate life-history information for other systems and species, generalizing the geographic component (i.e., structural river system) represents an important priority for future development.
[…]
