### 2025-01-21 #7475 @danStich
I've reviewed the final proof and noted two typos: 1) an extra "of" in the second paragraph under "Statement of need" and an extra "the" in the third paragraph of the same section. I updated and pushed this correction to the repository.
