### 2025-01-21 #7475 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. I see "r" and "bayesian" and I think those should be capitalized but please check throughly for others.
