### 2025-01-21 dd00191
reference capitalization for openjournals/joss-reviews#7475
