+++ paper.bib
@@ -104,7 +104,7 @@
 }
 
 @Manual{shiny,
-  title = {shiny: Web Application Framework for R},
+  title = {{shiny}: Web Application Framework for {R}},
   author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
   year = {2024},
   note = {R package version 1.10.0},
@@ -169,5 +169,5 @@
   volume = {7}, number = {71}, 
   pages = {3444}, 
   author = {Richard A. Erickson and Daniel S. Stich and Jillian L. Hebert}, 
-  title = {fishStan: Hierarchical Bayesian models for fisheries}, 
+  title = {{fishStan}: Hierarchical {Bayesian} models for fisheries}, 
   journal = {Journal of Open Source Software} }  
