### 2025-01-21 #7475 @danStich
Thank you. I checked all references and I noticed a couple other instances for R package names. It looks like these were references we added during review. New .bib file pushed to repository.
